+++ packages/react-native-reanimated/package.json
@@ -158,8 +158,8 @@
     "./lib/module/layoutReanimation/animationsManager.js",
     "./src/core.ts",
     "./lib/module/core.js",
-    "./src/initializers.ts",
-    "./lib/module/initializers.js",
+    "./src/initializers.native.ts",
+    "./lib/module/initializers.native.js",
     "./src/index.ts",
     "./lib/module/index.js"
   ]
