+++ packages/react-native-reanimated/src/ConfigHelper.native.ts
@@ -5,11 +5,6 @@ import { runOnUISync } from 'react-native-worklets';
 import type { LoggerConfig } from './common';
 import { getLoggerConfig, updateLoggerConfig } from './common';
 
-export {
-  addWhitelistedNativeProps,
-  addWhitelistedUIProps,
-} from './commonConfigHelper';
-
 /**
  * Updates Reanimated logger config with the user-provided configuration. Will
  * affect Reanimated code executed after call to this function so it should be
