+++ packages/react-native-reanimated/src/ReducedMotionCommon.ts
