+++ packages/react-native-reanimated/src/mappersCommon.ts
@@ -1,0 +1,13 @@
+'use strict';
+
+import type { MapperOutputs, SharedValue } from './commonTypes';
+
+export type MapperExtractedInputs = SharedValue[];
+
+export type Mapper = {
+  id: number;
+  dirty: boolean;
+  worklet: () => void;
+  inputs: MapperExtractedInputs;
+  outputs?: MapperOutputs;
+};
