+++ packages/react-native-reanimated/src/mutablesCommon.ts
@@ -2,7 +2,7 @@
 
 import type { ShareableHost, Synchronizable } from 'react-native-worklets';
 
-import { IS_JEST, logger } from './common';
+import { logger } from './common';
 import type { Mutable } from './commonTypes';
 import { isFirstReactRender, isReactRendering } from './reactUtils';
 import { valueSetter } from './valueSetter';
@@ -29,9 +29,9 @@ export function checkInvalidWriteDuringRender() {
   }
 }
 
-type Listener<TValue> = (newValue: TValue) => void;
+export type Listener<TValue> = (newValue: TValue) => void;
 
-type PartialMutable<TValue> = Omit<Mutable<TValue>, 'get' | 'set'>;
+export type PartialMutable<TValue> = Omit<Mutable<TValue>, 'get' | 'set'>;
 
 /**
  * Adds `get` and `set` methods to the mutable object to handle access to
@@ -169,70 +169,6 @@ export function mutableHostDecorator<TValue>(
   addCompilerSafeGetAndSet(mutable);
 
   return mutable;
-}
-
-export function makeMutableWeb<TValue>(initial: TValue): Mutable<TValue> {
-  let value: TValue = initial;
-  const listeners = new Map<number, Listener<TValue>>();
-
-  const mutable: PartialMutable<TValue> = {
-    get value(): TValue {
-      checkInvalidReadDuringRender();
-      return value;
-    },
-    set value(newValue) {
-      checkInvalidWriteDuringRender();
-      valueSetter(mutable as Mutable<TValue>, newValue);
-    },
-
-    get _value(): TValue {
-      return value;
-    },
-    set _value(newValue: TValue) {
-      value = newValue;
-      listeners.forEach((listener) => {
-        listener(newValue);
-      });
-    },
-
-    modify: (modifier, forceUpdate = true) => {
-      valueSetter(
-        mutable as Mutable<TValue>,
-        modifier !== undefined ? modifier(mutable.value) : mutable.value,
-        forceUpdate
-      );
-    },
-    addListener: (id: number, listener: Listener<TValue>) => {
-      listeners.set(id, listener);
-    },
-    removeListener: (id: number) => {
-      listeners.delete(id);
-    },
-
-    _isReanimatedSharedValue: true,
-  };
-
-  // Hide `_value` from accidental enumeration.
-  Object.defineProperty(mutable, '_value', {
-    configurable: false,
-    enumerable: false,
-  });
-
-  addCompilerSafeGetAndSet(mutable);
-
-  if (IS_JEST) {
-    (mutable as JestMutable<TValue>).toJSON = () => mutableToJSON(value);
-  }
-
-  return mutable as Mutable<TValue>;
-}
-
-interface JestMutable<TValue> extends Mutable<TValue> {
-  toJSON: () => string;
-}
-
-function mutableToJSON<TValue>(value: TValue): string {
-  return JSON.stringify(value);
 }
 
 /** @deprecated Used only in `animationsManager.ts`. Don't use. */
